Refactor generic GMT commands to dedicated submodule
Anything that looks remotely reusable is now in anf.gmt.

Also removed a bunch of unfinished migrations that will need to be
 refactored.
