+++ anf/bin/web/usarray_deploy_map/anf/deploymentmap/constant.py
@@ -15,42 +15,23 @@ MAX_YEAR = datetime.date.today().year
 # it's not really a constant, but close enough.
 
 VALID_YEARS = range(START_YEAR, MAX_YEAR + 1)  # "..up to but not including stop."
-VALID_MONTHS = range(1, 13)
+VALID_MONTHS = range(
+    1, 13
+)  # python range() is weird, includes start number, but excludes stop number.
 
 
 WET_RGB = "202/255/255"
 
 INTERMEDIATE_FORMAT = "PS"
-"""The format for the GMT working copy. Normally Postscript."""
+"""The format for the GMT working copy. Must be Postscript for GMT.
+
+This gets down-cased and used as the suffix of the intermediate file."""
 
 DEFAULT_OUTPUT_FORMAT = "PNG"
-"""The default format for the final output image, normally Portable Network Graphic (PNG)"""
+"""The default format for the final output image, normally Portable Network Graphic (PNG).
 
-# Plot Media options
-DEFAULT_PS_PAGE_ORIENTATION = "portrait"
-DEFAULT_PS_PAGE_COLOR = "255/255/255"
-DEFAULT_PS_MEDIA = "a1"
-
-# Basemap annotation options
-DEFAULT_MAP_ANNOT_OFFSET_PRIMARY = "0.2c"
-DEFAULT_MAP_ANNOT_OFFSET_SECONDARY = "0.2c"
-DEFAULT_MAP_LABEL_OFFSET = "0.2c"
-
-# Basemap Layout options
-DEFAULT_MAP_FRAME_WIDTH = "0.2c"
-DEFAULT_MAP_SCALE_HEIGHT = "0.2c"
-DEFAULT_MAP_TICK_LENGTH = "0.2c"
-DEFAULT_X_AXIS_LENGTH = "25c"
-DEFAULT_Y_AXIS_LENGTH = "15c"
-DEFAULT_MAP_ORIGIN_X = "2.5c"
-DEFAULT_MAP_ORIGIN_Y = "2.5c"
-DEFAULT_MAP_LOGO_POS = "BL/-0.2c/-0.2c"
-DEFAULT_MAP_LINE_STEP = "0.025c"
-
-# Misc options
-DEFAULT_PROJ_LENGTH_UNIT = "inch"
-DIR_GSHHG = "/usr/share/gshhg-gmt-nc4"
-
+This gets down-cased and used as the suffix of the final output image.
+"""
 
 DEFAULT_SYMSIZE = "0.15"
 DEFAULT_USE_COLOR = True
@@ -69,34 +50,38 @@ SIZE__DEPLOY_TYPE__FILE_FORMATS = {
     "wario": {
         "seismic": {
             "intermediate_file_prefix": "deployment_history_map_{deploy_type!s}_{year:04d}_{month:02d}_{map_type}_{size}_",
-            "intermediate_file_suffix": ".{intermediateformat}",
+            "intermediate_file_suffix": ".{intermediate_format}",
             # final file name was "make it yourself" message in original.
             "final_file_prefix": "deploymap_{year:04d}_{month:02d}.{map_type}_{size}",
-            "final_file_suffix": ".{outputformat}",
+            "final_file_suffix": ".{output_format}",
         },
         "inframet": {
             "intermediate_file_prefix": "deployment_history_map_{deploy_type!s}_{year:04d}_{month:02d}_{map_type}_{size}_",
-            "intermediate_file_suffix": ".{intermediateformat}",
+            "intermediate_file_suffix": ".{intermediate_format}",
             # final file name was "make it yourself" message in original.
             "final_file_prefix": "deploymap_{deploy_type}_{year:04d}_{month:02d}.{map_type}_{size}",
-            "final_file_suffix": ".{outputformat}",
+            "final_file_suffix": ".{output_format}",
         },
     },
     "default": {
         "seismic": {
             "intermediate_file_prefix": "deployment_history_map_{deploy_type!s}_{year:04d}_{month:02d}_{map_type}_",
-            "intermediate_file_suffix": ".{intermediateformat}",
+            "intermediate_file_suffix": ".{intermediate_format}",
             "final_file_prefix": "deploymap_{year:04d}_{month:02d}.{map_type}",
-            "final_file_suffix": ".{outputformat}",
+            "final_file_suffix": ".{output_format}",
         },
         "inframet": {
             "intermediate_file_prefix": "deployment_history_map_{deploy_type!s}_{year:04d}_{month:02d}_{map_type!s}_",
-            "intermediate_file_suffix": ".{intermediateformat}",
+            "intermediate_file_suffix": ".{intermediate_format}",
             "final_file_prefix": "deploymap_{deploy_type}_{year:04d}_{month:02d}.{map_type}",
-            "final_file_suffix": ".{outputformat}",
+            "final_file_suffix": ".{output_format}",
         },
     },
 }
-"""Filename format strings organized by size, then map_type."""
+"""Filename format strings organized by size, then map_type.
 
-DEPLOY_TYPE_DECOM_RGB = {"seismic": "77/77/77", "inframet": "255/255/255"}
+Retrieve the relevant sub key with DeploymentMapMaker._get_map_filename_parts().
+"""
+
+DEPLOY_TYPE_DECOM_RGBS = {"seismic": "77/77/77", "inframet": "255/255/255"}
+"""Color values for decomissioned stations, by deployment type."""
