+++ anf/bin/web/usarray_deploy_map/anf/gmt/Makefile
@@ -1,0 +1,9 @@
+DATADIR=python/anf/gmt
+
+DATA = $(filter-out setup.py, $(wildcard *.py))
+
+CLEAN = *.pyc
+
+include $(ANFMAKE)
+
+DIRS =
