+++ anf/bin/web/usarray_deploy_map/anf/gmt/constant.py
@@ -1,0 +1,18 @@
+"""Constants for the anf.gmt modules and functions."""
+DEFAULT_PS_PAGE_ORIENTATION = "portrait"
+DEFAULT_PS_PAGE_COLOR = "255/255/255"
+DEFAULT_PS_MEDIA = "a1"
+DEFAULT_MAP_ANNOT_OFFSET_PRIMARY = "0.2c"
+DEFAULT_MAP_ANNOT_OFFSET_SECONDARY = "0.2c"
+DEFAULT_MAP_LABEL_OFFSET = "0.2c"
+DEFAULT_MAP_FRAME_WIDTH = "0.2c"
+DEFAULT_MAP_SCALE_HEIGHT = "0.2c"
+DEFAULT_MAP_TICK_LENGTH = "0.2c"
+DEFAULT_X_AXIS_LENGTH = "25c"
+DEFAULT_Y_AXIS_LENGTH = "15c"
+DEFAULT_MAP_ORIGIN_X = "2.5c"
+DEFAULT_MAP_ORIGIN_Y = "2.5c"
+DEFAULT_MAP_LOGO_POS = "BL/-0.2c/-0.2c"
+DEFAULT_MAP_LINE_STEP = "0.025c"
+DEFAULT_PROJ_LENGTH_UNIT = "inch"
+DIR_GSHHG = "/usr/share/gshhg-gmt-nc4"
